+++ .storybook/components/DesignTokens/Tier1/FontSizes.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { Grid, Section } from '../../../../src';
+import filterTokens from '../../../util/filterTokens';
+import { TokenSpecimen } from '../../TokenSpecimen/TokenSpecimen';
+
+export const FontSizes = () => (
+  <Section title="Font Sizes">
+    <Grid>
+      {filterTokens(`eds-font-size`).map(function (listItem) {
+        return (
+          <Grid.Item key={listItem.name}>
+            <TokenSpecimen
+              inlineStyles={{
+                fontSize: `var(${listItem.name})`,
+              }}
+              name={listItem.name}
+              value={listItem.value}
+              variant="typography-title"
+            />
+          </Grid.Item>
+        );
+      })}
+    </Grid>
+  </Section>
+);