+++ .storybook/components/DesignTokens/Tier1/FontWeights.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import { Grid, Section } from '../../../../src';
+import filterTokens from '../../../util/filterTokens';
+import { TokenSpecimen } from '../../TokenSpecimen/TokenSpecimen';
+
+export const FontWeights = () => (
+  <Section title="Font Weights">
+    <Grid>
+      {filterTokens(`eds-font-weight`).map(function (listItem) {
+        return (
+          <Grid.Item key={listItem.name}>
+            <TokenSpecimen
+              inlineStyles={{
+                fontWeight: `var(${listItem.name})`,
+              }}
+              name={listItem.name}
+              value={listItem.value}
+              variant="typography-title"
+            />
+          </Grid.Item>
+        );
+      })}
+    </Grid>
+  </Section>
+);