+++ .storybook/pages/AlongDemo/index.ts
@@ -1,1 +1,0 @@
-export { AlongDemo as default } from './AlongDemo';