+++ .storybook/pages/CoursePlannerEdit/ProjectCard/ProjectCard.tsx
@@ -2,15 +2,7 @@ import clsx from 'clsx';
 import type { ReactNode } from 'react';
 import React from 'react';
 
-import {
-  Card,
-  CardHeader,
-  Heading,
-  Icon,
-  Button,
-  NumberIcon,
-  ButtonDropdown,
-} from '../../../../src';
+import { Card, Heading, Icon, Menu, NumberIcon } from '../../../../src';
 import type { HeadingElement } from '../../../../src/components/Heading';
 
 import type { IconName } from '../../../../src/components/Icon';
@@ -23,19 +15,11 @@ export interface Props {
    */
   behavior?: 'draggable';
   /**
-   * Button dropdown items
+   * Menu items
    *
-   * If not declared, the button dropdown does not render.
+   * If not declared, the menu does not render.
    */
-  buttonDropdownItems?: ReactNode;
-  /**
-   * Determines type of clickable
-   * - default renders a dropdown menu to the bottom left of the button
-   * - **top-left** renders a dropdown menu to the top left of the button
-   * - **top-right** renders a dropdown menu to the top right of the button
-   * - **bottom-right** renders a dropdown menu to the bottom right of the button
-   */
-  buttonDropdownPosition?: 'top-left' | 'top-right' | 'bottom-right';
+  menuItems?: ReactNode;
   /**
    * CSS class names that can be appended to the component.
    */
@@ -82,8 +66,7 @@ export const ProjectCard = ({
   metaIconName,
   number,
   headingAs = 'h3',
-  buttonDropdownItems,
-  buttonDropdownPosition,
+  menuItems,
   numberAriaLabel,
   isDragging,
   ...other
@@ -106,7 +89,7 @@ export const ProjectCard = ({
       orientation="horizontal"
       {...other}
     >
-      <CardHeader className={styles['project-card__header']}>
+      <Card.Header className={styles['project-card__header']}>
         {number && numberAriaLabel && (
           <NumberIcon
             aria-label={numberAriaLabel}
@@ -115,12 +98,12 @@ export const ProjectCard = ({
             size="sm"
           />
         )}
-      </CardHeader>
+      </Card.Header>
       <Card.Body className={styles['project-card__body']}>
         <Heading
           as={headingAs}
           className={styles['project-card__title']}
-          size="body-sm"
+          size="title-xs"
           variant="neutral-strong"
         >
           {title}
@@ -139,24 +122,28 @@ export const ProjectCard = ({
           </div>
         )}
       </Card.Body>
-      {buttonDropdownItems && (
+      {menuItems && (
         <Card.Footer className={styles['project-card__footer']}>
-          <ButtonDropdown
-            dropdownMenuTrigger={
-              <Button
-                aria-label="Open project dropdown"
-                className={styles['project-card__menu-button']}
-                size="sm"
-                status="neutral"
-                variant="icon"
-              >
-                <Icon name="dots-vertical" purpose="decorative" />
-              </Button>
-            }
-            position={buttonDropdownPosition}
+          <Menu
+            modifiers={[
+              // Brings the Menu closer to the dots-vertical icon trigger button
+              {
+                name: 'offset',
+                options: { offset: [0, 0] },
+              },
+            ]}
+            placement="bottom-end"
           >
-            {buttonDropdownItems}
-          </ButtonDropdown>
+            <Menu.PlainButton>
+              <Icon
+                name="dots-vertical"
+                purpose="informative"
+                size="1.5rem"
+                title="Open project dropdown"
+              />
+            </Menu.PlainButton>
+            <Menu.Items>{menuItems}</Menu.Items>
+          </Menu>
         </Card.Footer>
       )}
     </Card>