+++ src/components/InputField/InputField.tsx
@@ -8,7 +8,7 @@ import type {
 } from '../../util/utility-types';
 import FieldNote from '../FieldNote';
 import Input from '../Input';
-import Label from '../Label';
+import InputLabel from '../InputLabel';
 import Text from '../Text';
 import styles from './InputField.module.css';
 
@@ -127,6 +127,7 @@ export type Props = React.InputHTMLAttributes<HTMLInputElement> & {
 
 type InputFieldType = ForwardedRefComponent<HTMLInputElement, Props> & {
   Input?: typeof Input;
+  Label?: typeof InputLabel;
 };
 
 /**
@@ -157,7 +158,16 @@ export const InputField: InputFieldType = forwardRef(
     const overlineClassName = clsx(
       styles['input-field__overline'],
       !label && styles['input-field__overline--no-label'],
-      disabled && styles['input-field__overline--disabled'],
+    );
+
+    const labelClassName = clsx(
+      styles['input-field__label'],
+      disabled && styles['input-field__label--disabled'],
+    );
+
+    const requiredTextClassName = clsx(
+      styles['input-field__required-text'],
+      disabled && styles['input-field__required-text--disabled'],
     );
 
     const inputBodyClassName = clsx(
@@ -177,9 +187,13 @@ export const InputField: InputFieldType = forwardRef(
       <div className={className}>
         {shouldRenderOverline && (
           <div className={overlineClassName}>
-            {label && <Label htmlFor={idVar} text={label} />}
+            {label && (
+              <InputLabel className={labelClassName} htmlFor={idVar}>
+                {label}
+              </InputLabel>
+            )}
             {required && (
-              <Text as="p" size="sm">
+              <Text as="span" className={requiredTextClassName} size="sm">
                 Required
               </Text>
             )}
@@ -217,5 +231,7 @@ export const InputField: InputFieldType = forwardRef(
     );
   },
 );
+
 InputField.displayName = 'InputField';
 InputField.Input = Input;
+InputField.Label = InputLabel;